+++ types/index.d.ts
@@ -882,6 +882,8 @@ declare module "components/Code/index" {
             className: PropTypes.Requireable<string>;
             /** Style for the CopyButton */
             copyButtonStyle: PropTypes.Requireable<string>;
+            /** Copy button displays `Copy` instead of icon */
+            copyButtonUsesTextLabel: PropTypes.Requireable<boolean>;
             /** Adds a copy button to code examples */
             hasCopyButton: PropTypes.Requireable<boolean>;
             /** Apply syntax highlighting to the code */
@@ -895,6 +897,9 @@ declare module "components/Code/index" {
             /** How the code should be displayed */
             type: PropTypes.Validator<string>;
         };
+        export namespace defaultProps {
+            export const copyButtonUsesTextLabel: boolean;
+        }
     }
     import React from "react";
     import PropTypes from "prop-types";
@@ -1461,6 +1466,8 @@ declare module "components/Dropdown/DropdownBlockLink/index" {
     export namespace DropdownBlockLink {
         export namespace defaultProps {
             export const isLink: boolean;
+            export function onMouseEnter(): void;
+            export function onMouseLeave(): void;
         }
         export namespace propTypes {
             export const children: PropTypes.Validator<string | number | boolean | {} | PropTypes.ReactElementLike | PropTypes.ReactNodeArray>;
@@ -1471,8 +1478,10 @@ declare module "components/Dropdown/DropdownBlockLink/index" {
             export const isMultiSelect: PropTypes.Requireable<boolean>;
             export const minWidth: PropTypes.Requireable<string | number>;
             export const onClick: PropTypes.Validator<(...args: any[]) => any>;
-            export const onMouseEnter: PropTypes.Requireable<(...args: any[]) => any>;
-            export const onMouseLeave: PropTypes.Requireable<(...args: any[]) => any>;
+            const onMouseEnter_1: PropTypes.Requireable<(...args: any[]) => any>;
+            export { onMouseEnter_1 as onMouseEnter };
+            const onMouseLeave_1: PropTypes.Requireable<(...args: any[]) => any>;
+            export { onMouseLeave_1 as onMouseLeave };
             export const testSection: PropTypes.Requireable<string>;
             export const trackId: PropTypes.Requireable<string>;
             export const value: PropTypes.Requireable<string>;
@@ -1786,20 +1795,15 @@ declare module "components/DockedFooter/index" {
         state: {
             isDocked: boolean;
             listenersSet: boolean;
-            dockedFooterNode: null;
-            parentNode: null;
             atBottom: boolean;
-            shouldDockByHeight: boolean;
         };
         dockedFooterRef: React.RefObject<any>;
-        storeThisNode(): void;
-        domNodesInState(): boolean;
-        setShouldDockByHeight(): void;
-        setAtBottom(): void;
-        setEventListeners(): void;
-        throttle(delay: any, fn: any): (...args: any[]) => any;
-        componentDidMount(): void;
         componentDidUpdate(prevProps: any, prevState: any): void;
+        domNodesPresent: () => boolean;
+        setEventListeners: () => void;
+        throttle: (delay: any, fn: any) => (...args: any[]) => any;
+        setIsDocked: (callback: any) => void;
+        setAtBottom: () => void;
         componentWillUnmount(): void;
         render(): JSX.Element;
     }
@@ -1809,7 +1813,9 @@ declare module "components/DockedFooter/index" {
             export const includesMargin: PropTypes.Requireable<boolean>;
             export const leftGroup: PropTypes.Requireable<(PropTypes.ReactElementLike | null | undefined)[]>;
             export const rightGroup: PropTypes.Requireable<(PropTypes.ReactElementLike | null | undefined)[]>;
-            export const scrollRef: PropTypes.Requireable<PropTypes.ReactNodeLike>;
+            export const scrollRef: PropTypes.Requireable<((...args: any[]) => any) | PropTypes.InferProps<{
+                current: PropTypes.Requireable<PropTypes.ReactComponentLike>;
+            }>>;
             export const testSection: PropTypes.Validator<string>;
         }
         export namespace defaultProps {